sql,*: move GenerateUniqueInt out of `builtins`
This change is move-only change that moves GenerateUniqueInt
and some of its helpers and constants from `pkg/sql/sem/builtins`
to `pkg/util/unique`. This was primarily motivated by a need to
break the dependency of `pkg/jobs` on `pkg/sql/sem/builtins` so that
in the future `crdb_internal` builtins can use packages that might
import `pkg/jobs`.

Release note: None
