+++ pkg/sql/importer/BUILD.bazel
@@ -87,7 +87,6 @@ go_library(
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
         "//pkg/sql/rowexec",
-        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/catid",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
@@ -117,6 +116,7 @@ go_library(
         "//pkg/util/timeutil",
         "//pkg/util/timeutil/pgdate",
         "//pkg/util/tracing",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "//pkg/workload",
         "@com_github_cockroachdb_apd_v3//:apd",
