+++ pkg/sql/row/BUILD.bazel
@@ -60,6 +60,7 @@ go_library(
         "//pkg/sql/rowenc/valueside",
         "//pkg/sql/rowinfra",
         "//pkg/sql/scrub",
+        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/builtins/builtinconstants",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/transform",
