+++ pkg/sql/sem/builtins/BUILD.bazel
@@ -120,7 +120,6 @@ go_library(
         "//pkg/util/randutil",
         "//pkg/util/rangedesc",
         "//pkg/util/ring",
-        "//pkg/util/syncutil",
         "//pkg/util/timeofday",
         "//pkg/util/timetz",
         "//pkg/util/timeutil",
@@ -132,6 +131,7 @@ go_library(
         "//pkg/util/tsearch",
         "//pkg/util/ulid",
         "//pkg/util/unaccent",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
