+++ pkg/util/unique/unique_test.go
@@ -15,6 +15,8 @@ import (
 	"reflect"
 	"strconv"
 	"testing"
+
+	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 )
 
 func TestUniquifyByteSlices(t *testing.T) {
@@ -160,5 +162,25 @@ func TestUniquifyAcrossSlices(t *testing.T) {
 				t.Errorf("expected %v, got %v", tt.expectedRight, right)
 			}
 		})
+	}
+}
+
+// TestGenerateUniqueIDOrder verifies the expected ordering of
+// GenerateUniqueID.
+func TestGenerateUniqueIDOrder(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	tests := []uint64{
+		GenerateUniqueID(0, 0),
+		GenerateUniqueID(1, 0),
+		GenerateUniqueID(2<<15, 0),
+		GenerateUniqueID(0, 1),
+		GenerateUniqueID(0, 10000),
+		GenerateUniqueInt(0),
+	}
+	prev := tests[0]
+	for _, tc := range tests[1:] {
+		if tc <= prev {
+			t.Fatalf("%d > %d", tc, prev)
+		}
 	}
 }
